pyproject.toml
[tool.poetry]
package-mode = false
name = "jamming-calculator"
version = "0.1.0"
description = "Calculator to determine appropriate parameters for The Banshee's jamming mission."
license = "MIT"
authors = [
    "Matthew Mehrtens <12023414+mcmehrtens@users.noreply.github.com>",
]
readme = "README.md"
repository = "https://github.com/StratoShield-AERE-4610-Fall-2024/Jamming-Calculator"

[tool.poetry.dependencies]
python = "^3.13"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
